+++ scripts/libfile_examples.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+set -e
+
+. liblog.sh
+. libfile.sh
+
+fexist /etc/os-release
+ftype